refactor: Column 컴포넌트를 컴포지션 패턴으로 수정
